Working LoaderManagerFragment and initial implementation of ManagedList
MainActivity still does not use these in a useful way, but a LoaderManagerFragment is initialized and used to load MediaStore.Images.Media.EXTERNAL_CONTENT_URI.  I had to implement a pendingLoad list to process LoaderManagerFragment.load() calls that are made before onAttach() is called because Loader initialization requires the getActivity() call which is not available until after onAttach().  LoaderManagerFragment.load() still returns the id of the loader to be initialized so the Main thread can continue.  LoaderManagerFragment will need to verify the load is finished before providing any data requested of it.

ManagedList initial implementation is complete but untested.  All methods that modify the list are overridden to throw UnsupportedOperationException in favor of implementations that verify that a registered manager is making the modification request and subsequently notify all registered observers of modification.