Abandoned managed data set impementation.  Added "DISTINCT "
Implementation of managed data sets was becomming overly complicated.  After more research on Loader and SQLite I found adding the "DISTINCT" flag was the easiest way to get a list of unique BUCKET_IDs.  The Loader framework does not support this directly, however I found that adding the String "DISTINCT " to the first column in the projection array accomplished the same result.  MainActivity will support callbacks from LoaderManagerFragment to handle onLoadFinished and onLoaderReset calls.

QuickBarFragment now has a helper class called Item, a wrapper for data used for display.  MainActivity will be in charge of giving QuickBarFragment a List<Item> to display.  It will construct that list from data received in the LoaderManagerFragment callbacks.  A similar implementation should be used for DirectoryFragment.

I also renamed some classes and removed unused ones to clean up the environment.