[Working] Corrected one error after merge.
We need to set up drag and drop functionality for SmallQuickBarFragment so we can drag a directory onto the current item location.  Also I think we should implement a ViewTag enum to index tags attached to views to help process drag and drop.  For instance:

enum ViewTags{
   VIEW_HOLDER, ITEM_TYPE,...
}

enum ItemTypes{
   IMAGE, DIRECTORY,...
}

Then in onStartDrag we could add a tag to the view being dragged:
v.setTag(ViewTags.ITEM_TYPE, ItemTypes.DIRECTORY);

Then when the view is dropped we can see the item type and do selective processing.  But I'm not sure if there would be a better way to accomplish the same.