[Working] Sliding Pane only slides from the designated area.
At the moment, the designated area is the blueish bar at the left of the sliding pane.

Also, I renamed that bar. It was misspelled.  It was conentPaneDragArea... (missing a "t" in content). It is now contentPaneDragArea... (not missing the "t")

For the slidingpane, I created a class called SlidingPane that extends SlidingPaneLayout, and used that instead. For the sliding functionality, I override "onInterceptTouchEvent" to check if the touch is within the designated view.