[Working] Image drop on target now calls back to directoryfragment to move image.
Very strange bug in processAllBucketsLoadFinished.  I tried to do an initial load of the Camera bucket content.  If you look at the function, the if statement in it is not evaluating to true.  But I have set breakpoints and checked the values and for some reason Camera == Camera is not evaluating to true????  They even have the exact same hash code.