[Working] Version of Swipe
Left and Right Swipes are detected.
If you swipe too far to the left - past the slidingpane slide bar, it crashes and i don't know why.

But, this is just a test. We want swipe to be done like it is in the task manager, and I will try to figure that out.