[Working] Dragging image to info bubble opens menu
Added an info bubble above the menu button.
When you drag an image to it, a dialog (menu) opens up. We can choose what we want to do here.

Right now it has a spot for name, date, album, and description, and two buttons.

Also I put a colored horizontal bar at the top of this dialog box. Let me know what you think about that.